account: split key editor components
